% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_diffr.R
\name{qp_diffr}
\alias{qp_diffr}
\title{Show difference between files}
\usage{
qp_diffr(result, txt_dir)
}
\arguments{
\item{result}{One row of the results from `qp_which()`.}

\item{txt_dir}{The directory containing the extracted txt files}
}
\description{
Show difference between files
}
