% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{qp_extract_file}
\alias{qp_extract_file}
\title{Extract code a quarto file}
\usage{
qp_extract_file(file)
}
\arguments{
\item{file}{html file to process}
}
\description{
Extract code a quarto file
}
\examples{
f <- system.file("extdata/q3.html", package = "qpcontrol")
qp_extract_file(f) |>
  cat()
}
