% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{qp_pairwise}
\alias{qp_pairwise}
\title{Pairwise comparison of files}
\usage{
qp_pairwise(txt_dir)
}
\arguments{
\item{txt_dir}{directory where the extracted code is}
}
\description{
Pairwise comparison of files
}
