% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_report.R
\name{qp_report}
\alias{qp_report}
\title{Make a report of text similarity}
\usage{
qp_report(html_dir, threshold = 0.7, ...)
}
\arguments{
\item{html_dir}{Directory contraining html files}

\item{threshold}{Consider dissimilarities below this threshold}

\item{\dots}{extra arguments for quarto_render}
}
\description{
Make a report of text similarity
}
