% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{qp_rescue}
\alias{qp_rescue}
\title{Rescue .qmd files}
\usage{
qp_rescue(html_dir, txt_dir)
}
\arguments{
\item{html_dir}{Directory with html files to process.}

\item{txt_dir}{Directory with processed qmd files}
}
\description{
Some students will submit their qmd file, either because they cannot render it,
or because they just submit the wrong file.
 This function copies these files into directory with the extracted qmd files,
 but ignores qmd files when the student has also submitted an html file.
}
