% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{qp_unexected_filetype}
\alias{qp_unexected_filetype}
\title{Alert for unexpected filetypes}
\usage{
qp_unexected_filetype(html_dir)
}
\arguments{
\item{html_dir}{Directory with html files to process.}
}
\description{
Alert for unexpected filetypes
}
