% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qp_which.R
\name{qp_which}
\alias{qp_which}
\title{Which distances are below the threshold}
\usage{
qp_which(dist, thresh)
}
\arguments{
\item{dist}{distance matrix}

\item{thresh}{threshold distance}
}
\description{
Which distances are below the threshold
}
