% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_code.R
\name{qp_extract}
\alias{qp_extract}
\title{Extract code from quarto files}
\usage{
qp_extract(html_dir, txt_dir)
}
\arguments{
\item{html_dir}{Directory with html files}

\item{txt_dir}{Directory to fill with extracted code.}
}
\description{
Extract code from quarto files
}
\examples{
d <- system.file("extdata", package = "qpcontrol")
tmp <- tempdir()
qp_extract(html_dir = d, txt_dir = tmp)

}
